requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Werkzeug==2.3.7
reportlab==4.0.4
PyMySQL==1.1.0
python-dotenv==1.0.0
gunicorn
mysql-connector-python
scrypt